+++ src/design-system/forja/src/icons/index.ts
@@ -1,0 +1,2 @@
+export * from "./ForjaIcon";
+export * from "./ForjaCustomIcons";